Enhance QR code scanning functionality: Introduce a modal for scanning QR codes with improved user interface elements, including camera and file upload options. Implement robust error handling and validation for scanned QR data, ensuring better user feedback and experience. Update attendance processing logic to handle various scenarios, including expired and reused QR codes. Refactor QR code generation to include versioning for cache management.

+++ app/Livewire/QrScanner.php
@@ -286,7 +286,7 @@ class QrScanner extends Component
         } elseif ($minutes < 60) {
             return $minutes . ' ' . ($minutes === 1 ? 'minute' : 'minutes');
         } else {
-            $hours = floor($minutes / 60);
+            $hours = (int) floor($minutes / 60);
             $remainingMinutes = $minutes % 60;
             $hoursText = $hours . ' ' . ($hours === 1 ? 'hour' : 'hours');
             if ($remainingMinutes > 0) {
